+++ Makefile
@@ -4,7 +4,7 @@ YACC=bison
 
 CFLAGS=-Wall -g
 
-LINT_FILES=src/*.cpp src/*.h tests/**/*.cpp
+LINT_FILES=src/**/*.cpp src/**/*.h tests/**/*.cpp
 
 all: main
 
@@ -13,7 +13,6 @@ lint:
 
 fmt:
 	clang-format -i $(LINT_FILES)
-
 
 # unitテスト
 unit-test:
@@ -28,7 +27,7 @@ unit-test:
 		tests/unit/boundary/test_boundary.cpp \
 		tests/unit/arithmetic/test_arithmetic.cpp \
 		tests/unit/cross_type/test_cross_type.cpp \
-		src/eval.cpp
+		src/eval/eval.cpp
 	tests/unit/test_main
 
 # 結合テスト
@@ -39,11 +38,13 @@ integration-test: main
 		tests/integration/boundary/test_boundary.cpp \
 		tests/integration/assign/test_assign.cpp \
 		tests/integration/cross_type/test_cross_type.cpp \
-		src/eval.cpp
+		src/eval/eval.cpp
 	tests/integration/test_main
 
 # 両方まとめて実行
-test: unit-test integration-test
+test:
+	$(MAKE) unit-test
+	$(MAKE) integration-test
 
 src/lexer.c: src/lexer.l
 	$(LEX) -o src/lexer.c src/lexer.l
@@ -52,9 +53,16 @@ src/parser.c src/parser.h: src/parser.y
 	$(YACC) -d -o src/parser.c src/parser.y
 
 
-main: src/parser.c src/lexer.c src/main.cpp src/eval.cpp
-	$(CC) $(CFLAGS) -o main src/parser.c src/lexer.c src/main.cpp src/eval.cpp
+main: src/parser.c src/lexer.c src/main.cpp src/eval/eval.cpp src/ast/util.cpp
+	$(CC) $(CFLAGS) -I. -o main src/parser.c src/lexer.c src/main.cpp src/eval/eval.cpp src/ast/util.cpp
 
+# Cb→Cコード変換ツール
+.PHONY: cgen
+cgen:
+	$(CC) $(CFLAGS) -I. -o cgen_main cgen/cgen_main.cpp src/ast/util.cpp src/parser.c src/lexer.c
 
 clean:
-	rm -rf src/*.o src/*.c src/parser.h main tests/integration/test_main tests/unit/test_main tests/integration/*.dSYM tests/unit/*.dSYM *.dSYM
+	rm -rf src/*.o src/*.c src/parser.h main cgen_main \
+		tests/integration/test_main tests/unit/test_main \
+		tests/integration/*.dSYM tests/unit/*.dSYM *.dSYM \
+		cgen/*.dSYM
