+++ tests/unit/type/int/test_int.cpp
@@ -1,6 +1,6 @@
 #include <cassert>
 #include <iostream>
-#include "src/eval.h"
+#include "src/eval/eval.h"
 
 void test_unit_int() {
     // int型の範囲内
