+++ tests/unit/type/short/test_short.cpp
@@ -1,6 +1,6 @@
 #include <cassert>
 #include <iostream>
-#include "src/eval.h"
+#include "src/eval/eval.h"
 
 void test_unit_short() {
     // short型の範囲内
