+++ tests/unit/type/tiny/test_tiny.cpp
@@ -1,6 +1,6 @@
 #include <cassert>
 #include <iostream>
-#include "src/eval.h"
+#include "src/eval/eval.h"
 
 void test_unit_tiny() {
     // tiny型の範囲内
